requirements.txt
txt
selenium
python-dotenv
supabase
google-play-scraper